requirements.txt
flask
llama-cpp-python>=0.2.56
